Collector: split shared infrastructure by responsibility
Structure-only commit, and the style migration to the repo-standard
foo.rs + foo/ layout (collector was the last mod.rs holdout):

- collector.rs: facade — provider + infra module declarations and the
  re-exports that keep every caller addressing crate::collector::*.
- events.rs: FileEvents, the Keyed* wrappers, touch compression — the
  bundle the cache serializes.
- walk.rs: directory walking with mtime filtering and scan stats.
- cache.rs: CACHE_VERSION (and its bump log), file stamps, the
  versioned parse cache. A diff touching this file is a cache-
  semantics change by definition.
- merge.rs: cross-file merge and keyed dedup (merge_into, dedupe_into,
  earliest-timestamp rule, metadata fill).
- project.rs: cwd-to-project normalization, both OS variants.

Tests move with their subjects. Moving types between modules does not
change their bincode encoding, so CACHE_VERSION stays at 13. Also
corrects a doc carried from the original that claimed a TZ change
needs --no-cache — the cache records offset_seconds and rebuilds
automatically.

Second-model review: round 1 found only that stale doc (P3), round 2
clean. Same-instant A/B against a HEAD worktree build renders
byte-identical; original doc lines verified present by mechanical
comparison.
